BPF filter size is unsigned int

+++ bpf-filter.h
@@ -47,7 +47,7 @@ static const struct bpf_insn arp_bpf_filter [] = {
 	/* Otherwise, drop it. */
 	BPF_STMT(BPF_RET + BPF_K, 0),
 };
-static const size_t arp_bpf_filter_len =
+static const unsigned int arp_bpf_filter_len =
     sizeof(arp_bpf_filter) / sizeof(arp_bpf_filter[0]);
 
 
@@ -97,5 +97,5 @@ static const struct bpf_insn dhcp_bpf_filter [] = {
 	/* Otherwise, drop it. */
 	BPF_STMT(BPF_RET + BPF_K, 0),
 };
-static const size_t dhcp_bpf_filter_len =
+static const unsigned int dhcp_bpf_filter_len =
     sizeof(dhcp_bpf_filter) / sizeof(dhcp_bpf_filter[0]);
